split page.tsx
